Añadido primera version para borrar datos  brotherhood

+++ src/main/java/services/ChapterService.java
@@ -24,7 +24,6 @@ import security.Authority;
 import security.UserAccount;
 import domain.Actor;
 import domain.Area;
-import domain.Box;
 import domain.Brotherhood;
 import domain.Chapter;
 import domain.Configuration;
@@ -386,18 +385,22 @@ public class ChapterService {
 			final Collection<Proclaim> proclaims1 = this.proclaimService.findAll();
 			Assert.isTrue(!(proclaims1.contains(p)));
 		}
-		final Collection<Box> boxes = this.actorService.findBoxByActorId(chapter.getId());
-		for (final Box b : boxes) {
-			Assert.isTrue(b.getActor().getId() == chapter.getId());
-			this.boxService.delete(b);
-			final Collection<Box> boxes1 = this.boxService.findAll();
-			Assert.isTrue(!(boxes1.contains(b)));
-		}
+		//		final Collection<Box> boxes = this.actorService.findBoxByActorId(chapter.getId());
+		//		if (boxes != null)
+		//			for (final Box b : boxes) {
+		//
+		//				Assert.isTrue(b.getActor().getId() == chapter.getId());
+		//
+		//				this.boxService.delete1(b);
+		//
+		//			}
 		final Collection<SocialProfile> socialProfiles = this.socialProfileService.findProfileByActorId(chapter.getId());
-		for (final SocialProfile s : socialProfiles) {
-			Assert.isTrue(s.getActor().getId() == chapter.getId());
-			this.socialProfileService.delete(s);
-		}
+		if (socialProfiles != null)
+			for (final SocialProfile s : socialProfiles) {
+				Assert.isTrue(s.getActor().getId() == chapter.getId());
+				this.socialProfileService.delete(s);
+
+			}
 		this.chapterRepository.delete(chapter.getId());
 		final Collection<Actor> actors = this.actorService.findAll();
 		Assert.isTrue(!(actors.contains(chapter)));
